refactor: ProgramRecommend domain 정리
